python: Add new macros and function to py3compat.h
For Python 2 new macros redefine new PyCapsule functionality to
old PyCObject functionality for backwards compatibility.

In general, new macros enable us to use new PyCapsule_* functions
in the whole codebase without breaking backward compatibility with
PyCObject in Python 2. When the codebase will be ready to switch
to PyCapsule_*, the only thing will need to do is remove these
macros.

New function Capsule_FromVoidPtrAndContext() is here to handle
creating a new object from a void pointer with a description in
Python 2 (with CObject) and Python 3 (with Capsule).

Signed-off-by: Lumir Balhar <lbalhar@redhat.com>