python: Add PyCapsule-related macros and function to py3compat.h
For Python 2, new macros alias new PyCapsule functionality to
old PyCObject functionality for backwards compatibility.

In general, the new macros enable us to use new PyCapsule_* functions
in the whole codebase without breaking backward compatibility with
PyCObject in Python 2. When the codebase will be ready to switch
to PyCapsule_*, we'll only need to remove these macros.

The only exception is the new function Capsule_FromVoidPtrAndContext(),
which handles creating a new object from a void pointer
with a description. The CObject/Capsule APIs are too different
here to provide a simple macro.

This follows the strategy laid out by this comment in Python.pm:
   ## When we'll get support fpr Python3 we'll have to emulate
   ## PyCObject using PyCapsule and convert these functions to
   ## use PyCapsule.

Signed-off-by: Lumir Balhar <lbalhar@redhat.com>